@@ -3,6 +3,7 @@ using System;
 using System.CommandLine.Builder;
 using System.CommandLine.Invocation;
 using System.CommandLine.Parsing;
+using System.Diagnostics.CodeAnalysis;
 using System.Threading;
 using System.Threading.Tasks;
 using Upstream.CommandLine.Utilities;
@@ -20,7 +21,7 @@ namespace Upstream.CommandLine
                 .UseDefaults();
         }
 
-        public IServiceProvider ServiceProvider { get; private set; }
+        public IServiceProvider? ServiceProvider { get; private set; }
 
         public CommandLineApplication ConfigureServices(Action<IServiceCollection> configureServices)
         {
@@ -29,6 +30,7 @@ namespace Upstream.CommandLine
             return this;
         }
 
+        [MemberNotNull(nameof(ServiceProvider))]
         public Task<int> InvokeAsync(string[] args)
         {
             ServiceProvider = _services.BuildServiceProvider();
@@ -36,17 +38,21 @@ namespace Upstream.CommandLine
             return _builder.Build().InvokeAsync(args);
         }
 
-        public CommandLineApplication AddCommand<TAction, TOptions>(string name, string description = null)
+        public CommandLineApplication AddCommand<TAction, TOptions>(string name, string? description = null)
             where TAction : class, ICommandAction<TOptions>
             where TOptions : class
         {
-
             _services.AddScoped<TAction>();
 
             var command = new System.CommandLine.Command(name, description)
             {
                 Handler = CommandHandler.Create<TOptions, CancellationToken>((options, cancellationToken) =>
                 {
+                    if (ServiceProvider == null)
+                    {
+                        throw new InvalidOperationException("Command was invoked without building ServiceProvider");
+                    }
+
                     var action = ServiceProvider.GetRequiredService<TAction>();
 
                     return action.InvokeAsync(options, cancellationToken);
@@ -58,6 +64,13 @@ namespace Upstream.CommandLine
                 command.Add(symbol);
             }
 
+            _builder.AddCommand(command);
+
+            return this;
+        }
+
+        public CommandLineApplication AddCommand(System.CommandLine.Command command)
+        {
             _builder.AddCommand(command);
 
             return this;
